Distinguish from integration test Put common function in new file
