+++ code/go/0chain.net/blobbercore/handler/handler_share_test.go
@@ -1,3 +1,6 @@
+//go:build !integration
+// +build !integration
+
 package handler
 
 import (
